Use a subclass of Money (py-money) for all money operations.
All money operations are performed using Decimal under the hood, so we
avoid rounding problems. We subclassed the Money class to omit the
currency from its __repr__, so we can render it directly in the
templates. We are also less strict in the initialization amount for the
class, we accept arbitrary number of decimal places but round it to the
correct places during initialization.